requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
monai==1.3.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pillow==10.3.0
psutil==6.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scikit-learn-extra==0.3.0
scipy==1.13.1
segment-anything-hq==0.3
six==1.16.0
sympy==1.12.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==1.0.7
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.4
triton==2.1.0
typing_extensions==4.12.2
urllib3==2.2.2
